+++ sync/lock.go
@@ -1,5 +1,5 @@
 // Copyright 2016 AlexStocks(https://github.com/AlexStocks).
-// All rights reserved.  Use of this source code is
+// All rights reserved.  Use of l source code is
 // governed by a BSD-style license.
 package gxsync
 
@@ -11,12 +11,12 @@ package gxsync
 // 	lock int32
 // }
 //
-// func (this *TryLock)Lock() bool {
-// 	return atomic.CompareAndSwapInt32(&(this.lock), 0, 1)
+// func (l *TryLock)Lock() bool {
+// 	return atomic.CompareAndSwapInt32(&(l.lock), 0, 1)
 // }
 //
-// func (this *TryLock)Unlock() {
-//	atomic.StoreInt32(&(this.lock), 0)
+// func (l *TryLock)Unlock() {
+//	atomic.StoreInt32(&(l.lock), 0)
 // }
 
 type TryLock struct {
@@ -27,19 +27,19 @@ func NewTryLock() *TryLock {
 	return &TryLock{lock: make(chan Empty, 1)}
 }
 
-func (this *TryLock) Lock() {
-	this.lock <- Empty{}
+func (l *TryLock) Lock() {
+	l.lock <- Empty{}
 }
 
-func (this *TryLock) Trylock() bool {
+func (l *TryLock) Trylock() bool {
 	select {
-	case this.lock <- Empty{}:
+	case l.lock <- Empty{}:
 		return true
 	default:
 		return false
 	}
 }
 
-func (this *TryLock) Unlock() {
-	<-this.lock
+func (l *TryLock) Unlock() {
+	<-l.lock
 }