+++ sync/semaphore.go
@@ -1,5 +1,5 @@
 // Copyright 2016 AlexStocks(https://github.com/AlexStocks).
-// All rights reserved.  Use of this source code is
+// All rights reserved.  Use of s source code is
 // governed by a BSD-style license.
 package gxsync
 
@@ -17,24 +17,24 @@ func NewSemaphore(parallelNum int) *Semaphore {
 	return &Semaphore{lock: make(chan Empty, parallelNum)}
 }
 
-func (this *Semaphore) Post() {
-	<-this.lock
+func (s *Semaphore) Post() {
+	<-s.lock
 }
 
-// func (this *Semaphore) Release() {
-// 	var num = cap(this.lock) - len(this.lock)
+// func (s *Semaphore) Release() {
+// 	var num = cap(s.lock) - len(s.lock)
 // 	for i := 0; i < num; i++ {
-// 		<-this.lock
+// 		<-s.lock
 // 	}
 // }
 
-func (this *Semaphore) Wait() {
-	this.lock <- Empty{}
+func (s *Semaphore) Wait() {
+	s.lock <- Empty{}
 }
 
-func (this *Semaphore) TryWait() bool {
+func (s *Semaphore) TryWait() bool {
 	select {
-	case this.lock <- Empty{}:
+	case s.lock <- Empty{}:
 		return true
 	default:
 		return false
@@ -42,11 +42,11 @@ func (this *Semaphore) TryWait() bool {
 }
 
 // @timeout is in nanoseconds
-func (this *Semaphore) TimeWait(timeout int) bool {
+func (s *Semaphore) TimeWait(timeout int) bool {
 	select {
-	case this.lock <- Empty{}:
+	case s.lock <- Empty{}:
 		return true
-	case time.After(time.Duration(timeout * int(time.Nanosecond))):
+	case <-time.After(time.Duration(timeout * int(time.Nanosecond))):
 		return false
 	}
 }